% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhaolab3.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{data.frame}

\item{init_node}{numeric}
}
\description{
function name :dijkstra.This function take a data.frame and a
numeric variable as parameter.In data.frame,there should be 3 column named v1
v2,w separately.v1 and v2 and w make up a graph.

For example:
    if v1[1] = 1, v2[1] = 3,w[1] =4,
that means there is a edge linking V1 and V2,length is 2
    if v1[3] = 5, v2[3] = 8,w[3] =10,
that means there is a edge linking V5 and V8,length is 10

the function will return a vector which contains the shortest distance
from Vinit_node to each other points.(for something like v1 to v1 it's 0)
}
\details{
This is lab3 assignment from Wuhao Wang liuid:wuhwa469

Here recursive method is applied.First,form a matrix mat1,which
mat1[x,y] is the length of edge linking Vx and Vy.If no edge there,mat1[x,y]
will be filled by NA_real_.

There is another tool function to do recursive procedure,this function will
test every possible solution one by one and record the minimum solution.By
the way,in case of loop(like 1->3->1),there is a vector called "path" to
record visited point in case of  repeating.

reference [link url:](https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm)
}
\author{
Wuhao Wang
}
