% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhaolab3.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(x1, x2)
}
\arguments{
\item{x1, x2}{both are numeric}
}
\description{
function name:euclidean
            return the maximum common divisor of x1 and x2
}
\details{
This is lab3 assignment from Wuhao Wang liuid:wuhwa469

By testing numbers from 1 to sqrt(min(x1,x2)),find all 'common divisor

   If one of the parameter is negative,it will stop by calling stop().

   If 0 is applied,it will also stop by calling stop().

   Reference [link url:](https://en.wikipedia.org/wiki/Euclidean_algorithm)
}
\examples{
   x1<-4;x2<-6;euclidean(x1,x2)

   x1<-8;x2<-1;euclidean(x1,x2)

}
\author{
Wuhao Wang
}
