% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\description{
(v1,v2,w)means a graph.For example,as seen in column 1 (1,2,7)
    means there is edge linking V1 and V2 whose length is 7,as well as last 
    column (6,5,9) means there is edge linking V5 and V6 whose length is 9.
}
\references{
/url:https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
